Adds icon component and logic for getting icon related to weather description
