+++ .cache/11/c56d07b4918c7021fbd45901d6974b.json
@@ -1,0 +1,1 @@
+{"id":"components/Icons.js","dependencies":[{"name":"/Users/nolan/Documents/2019 - Projects/current-weather/.babelrc","includedInParent":true,"mtime":1575663687955},{"name":"/Users/nolan/Documents/2019 - Projects/current-weather/package.json","includedInParent":true,"mtime":1578671597257},{"name":"react","loc":{"line":1,"column":18},"parent":"/Users/nolan/Documents/2019 - Projects/current-weather/src/components/Icons.js","resolved":"/Users/nolan/Documents/2019 - Projects/current-weather/node_modules/react/index.js"},{"name":"@fortawesome/react-fontawesome","loc":{"line":2,"column":32},"parent":"/Users/nolan/Documents/2019 - Projects/current-weather/src/components/Icons.js","resolved":"/Users/nolan/Documents/2019 - Projects/current-weather/node_modules/@fortawesome/react-fontawesome/index.es.js"}],"generated":{"js":"\"use strict\";\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.Icons = void 0;\n\nvar _react = _interopRequireDefault(require(\"react\"));\n\nvar _reactFontawesome = require(\"@fortawesome/react-fontawesome\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nconst Icons = ({\n  currentIcon\n}) => {\n  return _react.default.createElement(_reactFontawesome.FontAwesomeIcon, {\n    icon: currentIcon,\n    size: \"10x\"\n  });\n};\n\nexports.Icons = Icons;"},"sourceMaps":{"js":{"mappings":[{"generated":{"line":8,"column":0},"source":"components/Icons.js","original":{"line":1,"column":0}},{"generated":{"line":10,"column":0},"source":"components/Icons.js","original":{"line":2,"column":0}},{"generated":{"line":14,"column":0},"source":"components/Icons.js","original":{"line":4,"column":7}},{"name":"Icons","generated":{"line":14,"column":6},"source":"components/Icons.js","original":{"line":4,"column":13}},{"generated":{"line":14,"column":11},"source":"components/Icons.js","original":{"line":4,"column":18}},{"generated":{"line":14,"column":14},"source":"components/Icons.js","original":{"line":4,"column":21}},{"generated":{"line":14,"column":15},"source":"components/Icons.js","original":{"line":4,"column":22}},{"name":"currentIcon","generated":{"line":15,"column":0},"source":"components/Icons.js","original":{"line":4,"column":23}},{"name":"currentIcon","generated":{"line":15,"column":2},"source":"components/Icons.js","original":{"line":4,"column":23}},{"generated":{"line":16,"column":0},"source":"components/Icons.js","original":{"line":4,"column":22}},{"generated":{"line":16,"column":1},"source":"components/Icons.js","original":{"line":4,"column":21}},{"generated":{"line":16,"column":6},"source":"components/Icons.js","original":{"line":4,"column":40}},{"generated":{"line":17,"column":0},"source":"components/Icons.js","original":{"line":6,"column":2}},{"generated":{"line":17,"column":9},"source":"components/Icons.js","original":{"line":7,"column":4}},{"generated":{"line":17,"column":38},"source":"components/Icons.js","original":{"line":7,"column":5}},{"generated":{"line":17,"column":71},"source":"components/Icons.js","original":{"line":7,"column":4}},{"generated":{"line":18,"column":0},"source":"components/Icons.js","original":{"line":7,"column":21}},{"generated":{"line":18,"column":4},"source":"components/Icons.js","original":{"line":7,"column":21}},{"generated":{"line":18,"column":8},"source":"components/Icons.js","original":{"line":7,"column":25}},{"name":"currentIcon","generated":{"line":18,"column":10},"source":"components/Icons.js","original":{"line":7,"column":27}},{"generated":{"line":18,"column":21},"source":"components/Icons.js","original":{"line":7,"column":4}},{"generated":{"line":19,"column":0},"source":"components/Icons.js","original":{"line":7,"column":40}},{"generated":{"line":19,"column":4},"source":"components/Icons.js","original":{"line":7,"column":40}},{"generated":{"line":19,"column":8},"source":"components/Icons.js","original":{"line":7,"column":44}},{"generated":{"line":19,"column":10},"source":"components/Icons.js","original":{"line":7,"column":45}},{"generated":{"line":20,"column":0},"source":"components/Icons.js","original":{"line":7,"column":4}},{"generated":{"line":20,"column":4},"source":"components/Icons.js","original":{"line":6,"column":2}},{"generated":{"line":21,"column":0},"source":"components/Icons.js","original":{"line":9,"column":1}},{"generated":{"line":21,"column":1},"source":"components/Icons.js","original":{"line":4,"column":7}}],"sources":{"components/Icons.js":"import React from \"react\";\nimport { FontAwesomeIcon } from '@fortawesome/react-fontawesome';\n\nexport const Icons = ({currentIcon}) => {\n\n  return (\n    <FontAwesomeIcon icon={currentIcon} size='10x' />\n  )\n}"},"lineCount":null}},"error":null,"hash":"49194c68bf53b753152160254b1c645a","cacheData":{"env":{}}}
