+++ .cache/de/1cc91a5e9f5236b08815f55cce3038.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/react-is/index.js","dependencies":[{"name":"/Users/nolan/Documents/2019 - Projects/current-weather/package.json","includedInParent":true,"mtime":1578671597257},{"name":"/Users/nolan/Documents/2019 - Projects/current-weather/node_modules/react-is/package.json","includedInParent":true,"mtime":1574374177307},{"name":"./cjs/react-is.development.js","loc":{"line":6,"column":27},"parent":"/Users/nolan/Documents/2019 - Projects/current-weather/node_modules/react-is/index.js","resolved":"/Users/nolan/Documents/2019 - Projects/current-weather/node_modules/react-is/cjs/react-is.development.js"}],"generated":{"js":"'use strict';\n\nif (\"development\" === 'production') {\n  module.exports = require('./cjs/react-is.production.min.js');\n} else {\n  module.exports = require('./cjs/react-is.development.js');\n}"},"sourceMaps":{"js":{"mappings":[{"generated":{"line":1,"column":0},"source":"../node_modules/react-is/index.js","original":{"line":1,"column":0}},{"generated":{"line":3,"column":0},"source":"../node_modules/react-is/index.js","original":{"line":3,"column":0}},{"generated":{"line":3,"column":4},"source":"../node_modules/react-is/index.js","original":{"line":3,"column":4}},{"generated":{"line":3,"column":22},"source":"../node_modules/react-is/index.js","original":{"line":3,"column":29}},{"generated":{"line":3,"column":34},"source":"../node_modules/react-is/index.js","original":{"line":3,"column":0}},{"generated":{"line":3,"column":36},"source":"../node_modules/react-is/index.js","original":{"line":3,"column":43}},{"name":"module","generated":{"line":4,"column":0},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":2}},{"name":"module","generated":{"line":4,"column":2},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":2}},{"generated":{"line":4,"column":8},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":8}},{"name":"exports","generated":{"line":4,"column":9},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":9}},{"generated":{"line":4,"column":16},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":2}},{"name":"require","generated":{"line":4,"column":19},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":19}},{"generated":{"line":4,"column":26},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":26}},{"generated":{"line":4,"column":27},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":27}},{"generated":{"line":4,"column":61},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":26}},{"generated":{"line":4,"column":62},"source":"../node_modules/react-is/index.js","original":{"line":4,"column":2}},{"generated":{"line":5,"column":0},"source":"../node_modules/react-is/index.js","original":{"line":5,"column":1}},{"generated":{"line":5,"column":1},"source":"../node_modules/react-is/index.js","original":{"line":3,"column":0}},{"generated":{"line":5,"column":7},"source":"../node_modules/react-is/index.js","original":{"line":5,"column":7}},{"name":"module","generated":{"line":6,"column":0},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":2}},{"name":"module","generated":{"line":6,"column":2},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":2}},{"generated":{"line":6,"column":8},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":8}},{"name":"exports","generated":{"line":6,"column":9},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":9}},{"generated":{"line":6,"column":16},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":2}},{"name":"require","generated":{"line":6,"column":19},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":19}},{"generated":{"line":6,"column":26},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":26}},{"generated":{"line":6,"column":27},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":27}},{"generated":{"line":6,"column":58},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":26}},{"generated":{"line":6,"column":59},"source":"../node_modules/react-is/index.js","original":{"line":6,"column":2}},{"generated":{"line":7,"column":0},"source":"../node_modules/react-is/index.js","original":{"line":7,"column":1}}],"sources":{"../node_modules/react-is/index.js":"'use strict';\n\nif (process.env.NODE_ENV === 'production') {\n  module.exports = require('./cjs/react-is.production.min.js');\n} else {\n  module.exports = require('./cjs/react-is.development.js');\n}\n"},"lineCount":null}},"error":null,"hash":"edbbbf037373bc4fb739a7a65eaaf1ad","cacheData":{"env":{"NODE_ENV":"development"}}}
